fix the api gateway issue in nginx ,and did some changes in backend and frontend,

+++ microservices-app/product-service/server.js
@@ -1,11 +1,27 @@
 require("dotenv").config();
 const express = require("express");
 const cors = require("cors");
+const jwt = require("jsonwebtoken");
 const Product = require("./models/Product");
 
 const app = express();
 app.use(cors());
 app.use(express.json());
+
+function verifyToken(req, res, next) {
+    const token = req.headers.authorization?.split(" ")[1];
+
+    if (!token) {
+        return res.status(401).json({ message: "No token" });
+    }
+
+    try {
+        req.user = jwt.verify(token, process.env.JWT_SECRET || "supersecretkey");
+        next();
+    } catch {
+        res.status(403).json({ message: "Invalid token" });
+    }
+}
 
 app.get("/", (req, res) => {
     res.json({
@@ -24,20 +40,18 @@ app.get("/health", (req, res) => {
     });
 });
 
-app.post("/products", async (req, res) => {
-    const userId = req.headers["x-user-id"] || "anonymous";
-    const product = await Product.create({ userId, ...req.body });
+app.post("/products", verifyToken, async (req, res) => {
+    const product = await Product.create({ userId: req.user.id, ...req.body });
     res.json(product);
 });
 
-app.get("/products", async (req, res) => {
-    const userId = req.headers["x-user-id"] || "anonymous";
-    const products = await Product.find({ userId });
+app.get("/products", verifyToken, async (req, res) => {
+    const products = await Product.find({ userId: req.user.id });
     res.json(products);
 });
 
-app.delete("/products/:id", async (req, res) => {
-    await Product.deleteOne({ _id: req.params.id });
+app.delete("/products/:id", verifyToken, async (req, res) => {
+    await Product.deleteOne({ _id: req.params.id, userId: req.user.id });
     res.json({ message: "Deleted" });
 });
